requirements.txt
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0
PyGithub==1.59.1
